Move StrongTypes schema classification to core
